requirements.txt
asgiref
autopep8
Babel
cssselect
Django
django-environ
django-filter
django-formtools
django-model-utils
django-notifications-hq
django-phonenumber-field
django-selenium-login
django-tables2
django-allauth
djangorestframework
html5lib
jsonfield
lxml
mysqlclient
phonenumbers
Pillow
pycodestyle
PyPDF2
pyquery
pytz
reportlab
selenium
six
sqlparse
swapper
urllib3
webencodings
xhtml2pdf
black
